refactor/#6: sns event 도메인 DDD 패키지 구조로 변경
- application, domain, infrastructure, presentation 구조로 변경
