+++ src/main/java/com/haru/api/domain/moodTracker/controller/MoodTrackerController.java
@@ -5,7 +5,7 @@ import com.haru.api.domain.moodTracker.dto.MoodTrackerResponseDTO;
 import com.haru.api.domain.moodTracker.entity.MoodTracker;
 import com.haru.api.domain.moodTracker.service.MoodTrackerCommandService;
 import com.haru.api.domain.moodTracker.service.MoodTrackerQueryService;
-import com.haru.api.domain.snsEvent.entity.enums.Format;
+import com.haru.api.snsEvent.domain.enums.Format;
 import com.haru.api.user.domain.User;
 import com.haru.api.workspace.domain.Workspace;
 import com.haru.api.global.annotation.AuthMoodTracker;
