+++ src/main/java/com/haru/api/domain/moodTracker/service/MoodTrackerCommandService.java
@@ -3,7 +3,7 @@ package com.haru.api.domain.moodTracker.service;
 import com.haru.api.domain.moodTracker.dto.MoodTrackerRequestDTO;
 import com.haru.api.domain.moodTracker.dto.MoodTrackerResponseDTO;
 import com.haru.api.domain.moodTracker.entity.MoodTracker;
-import com.haru.api.domain.snsEvent.entity.enums.Format;
+import com.haru.api.snsEvent.domain.enums.Format;
 import com.haru.api.user.domain.User;
 import com.haru.api.workspace.domain.Workspace;
