+++ src/main/java/com/haru/api/domain/moodTracker/service/MoodTrackerReportService.java
@@ -1,8 +1,7 @@
 package com.haru.api.domain.moodTracker.service;
 
-import com.haru.api.domain.moodTracker.dto.MoodTrackerResponseDTO;
 import com.haru.api.domain.moodTracker.entity.MoodTracker;
-import com.haru.api.domain.snsEvent.entity.enums.Format;
+import com.haru.api.snsEvent.domain.enums.Format;
 
 public interface MoodTrackerReportService {
     void generateReport(Long moodTrackerId);
