--- src/main/java/com/haru/api/domain/snsEvent/entity/enums/Format.java
@@ -1,5 +1,0 @@
-package com.haru.api.domain.snsEvent.entity.enums;
-
-public enum Format {
-    PDF, DOCX
-}
