--- src/main/java/com/haru/api/domain/snsEvent/entity/enums/ListType.java
@@ -1,5 +1,0 @@
-package com.haru.api.domain.snsEvent.entity.enums;
-
-public enum ListType {
-    PARTICIPANT, WINNER
-}
