+++ src/main/java/com/haru/api/global/documentFinder/SnsEventFinder.java
@@ -1,7 +1,7 @@
 package com.haru.api.global.documentFinder;
 
 import com.haru.api.workspace.domain.enums.DocumentType;
-import com.haru.api.domain.snsEvent.repository.SnsEventRepository;
+import com.haru.api.snsEvent.infrastructure.SnsEventRepository;
 import com.haru.api.global.apiPayload.code.status.ErrorStatus;
 import com.haru.api.global.apiPayload.exception.handler.SnsEventHandler;
 import lombok.RequiredArgsConstructor;
