+++ src/main/java/com/haru/api/global/interceptor/DocumentMemberAuthInterceptor.java
@@ -3,7 +3,7 @@ package com.haru.api.global.interceptor;
 import com.haru.api.workspace.domain.enums.DocumentType;
 import com.haru.api.meeting.infrastructure.MeetingRepository;
 import com.haru.api.domain.moodTracker.repository.MoodTrackerRepository;
-import com.haru.api.domain.snsEvent.repository.SnsEventRepository;
+import com.haru.api.snsEvent.infrastructure.SnsEventRepository;
 import com.haru.api.user.domain.User;
 import com.haru.api.user.infrastructure.UserRepository;
 import com.haru.api.infra.security.jwt.SecurityUtil;
