+++ src/main/java/com/haru/api/meeting/application/port/in/MeetingQueryUseCase.java
@@ -2,7 +2,7 @@ package com.haru.api.meeting.application.port.in;
 
 import com.haru.api.meeting.presentation.dto.MeetingResponseDTO;
 import com.haru.api.meeting.domain.Meeting;
-import com.haru.api.domain.snsEvent.entity.enums.Format;
+import com.haru.api.snsEvent.domain.enums.Format;
 import com.haru.api.user.domain.User;
 import com.haru.api.workspace.domain.Workspace;
