+++ src/main/java/com/haru/api/snsEvent/domain/Participant.java
@@ -1,4 +1,4 @@
-package com.haru.api.domain.snsEvent.entity;
+package com.haru.api.snsEvent.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -6,14 +6,14 @@ import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.DynamicUpdate;
 
 @Entity
-@Table(name = "winners")
+@Table(name = "participants")
 @Getter
 @DynamicUpdate
 @DynamicInsert
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-public class Winner {
+public class Participant {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -29,8 +29,8 @@ public class Winner {
     public void setSnsEvent(SnsEvent snsEvent) {
         this.snsEvent = snsEvent;
         if (this.snsEvent != null) {
-            snsEvent.getWinnerList().remove(this);
+            snsEvent.getParticipantList().remove(this);
         }
-        this.snsEvent.getWinnerList().add(this);
+        this.snsEvent.getParticipantList().add(this);
     }
 }
