+++ src/main/java/com/haru/api/snsEvent/domain/enums/Format.java
@@ -1,0 +1,5 @@
+package com.haru.api.snsEvent.domain.enums;
+
+public enum Format {
+    PDF, DOCX
+}
