+++ src/main/java/com/haru/api/snsEvent/domain/enums/ListType.java
@@ -1,0 +1,5 @@
+package com.haru.api.snsEvent.domain.enums;
+
+public enum ListType {
+    PARTICIPANT, WINNER
+}
