+++ src/main/java/com/haru/api/snsEvent/infrastructure/ParticipantRepository.java
@@ -1,7 +1,7 @@
-package com.haru.api.domain.snsEvent.repository;
+package com.haru.api.snsEvent.infrastructure;
 
-import com.haru.api.domain.snsEvent.entity.Participant;
-import com.haru.api.domain.snsEvent.entity.SnsEvent;
+import com.haru.api.snsEvent.domain.Participant;
+import com.haru.api.snsEvent.domain.SnsEvent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
