+++ src/main/java/com/haru/api/workspace/application/converter/WorkspaceConverter.java
@@ -4,7 +4,7 @@ import com.haru.api.workspace.domain.UserDocumentLastOpened;
 import com.haru.api.workspace.domain.enums.DocumentType;
 import com.haru.api.meeting.domain.Meeting;
 import com.haru.api.domain.moodTracker.entity.MoodTracker;
-import com.haru.api.domain.snsEvent.entity.SnsEvent;
+import com.haru.api.snsEvent.domain.SnsEvent;
 import com.haru.api.user.presentation.dto.UserResponseDTO;
 import com.haru.api.workspace.presentation.dto.WorkspaceResponseDTO;
 import com.haru.api.workspace.domain.Workspace;
