+++ src/main/java/com/haru/api/workspace/application/service/WorkspaceQueryUseCaseImpl.java
@@ -7,8 +7,8 @@ import com.haru.api.meeting.domain.Meeting;
 import com.haru.api.meeting.infrastructure.MeetingRepository;
 import com.haru.api.domain.moodTracker.entity.MoodTracker;
 import com.haru.api.domain.moodTracker.repository.MoodTrackerRepository;
-import com.haru.api.domain.snsEvent.entity.SnsEvent;
-import com.haru.api.domain.snsEvent.repository.SnsEventRepository;
+import com.haru.api.snsEvent.domain.SnsEvent;
+import com.haru.api.snsEvent.infrastructure.SnsEventRepository;
 import com.haru.api.user.application.converter.UserConverter;
 import com.haru.api.user.presentation.dto.UserResponseDTO;
 import com.haru.api.user.domain.User;
